+++ envs/src/raylibc/scripts/common/dlib.sh
@@ -1,3 +1,7 @@
 #!/bin/sh
 
-cc -Og -g -o target/hot/libhot.so radiance/extern.c target/hot/libraylib.so.??? -fPIC -shared -lm
+cc -Og -g \
+   -o target/hot/libhot.so \
+   radiance/extern.c \
+   target/hot/libraylib.so.??? \
+   -fPIC -shared -lm