+++ envs/src/raylibc/scripts/common/hotcore.sh
@@ -1,3 +1,8 @@
 #!/bin/sh
 
-cc -Og -g -o target/hot/main src/main.c target/hot/libraylib.so.??? -lltdl -ldl -lm -DRADIANCE_HOT
+cc -Og -g \
+   -o target/hot/main \
+   src/main.c \
+   target/hot/libraylib.so.??? \
+   -lltdl -ldl -lm \
+   -DRADIANCE_HOT