+++ envs/src/raylibcpp/radiance/lib.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+// #define RADIANCE_HOT
+
+#ifdef RADIANCE_HOT
+    #include "hot.hpp"
+#else
+    #include "static.hpp"
+#endif