+++ envs/src/raylibcpp/scripts/build.sh
@@ -1,0 +1,9 @@
+#!/bin/sh
+
+mkdir -p target/linux
+# cc -Ofast -o target/linux/release src/main.c -lraylib -lm
+g++ -Ofast \
+   -o target/linux/release \
+   src/main.cpp \
+   -Wl,-rpath,target/pkg/raylib/build/raylib \
+   -lraylib -lm