+++ envs/src/raylibcpp/scripts/common/dlib.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+
+g++ -Og -g \
+    -o target/hot/libhot.so \
+    radiance/extern.cpp \
+    target/hot/libraylib.so.??? \
+    -fPIC -shared -lm