+++ envs/src/raylibcpp/scripts/common/hotcore.sh
@@ -1,0 +1,8 @@
+#!/bin/sh
+
+g++ -Og -g \
+    -o target/hot/main \
+    src/main.cpp \
+    target/hot/libraylib.so.??? \
+    -lltdl -ldl -lm \
+    -DRADIANCE_HOT