+++ envs/src/raylibcpp/scripts/docs.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+xdg-open https://imsumire.github.io/radiance/docs